Better flow in ReadShips, hacky imagePath implementation
ReadShips modified so that there is no public array of ship record
lines. Rather, readShipRecord() creates an array which is passed
around. All methods made private except printTempShip and makeShip
such that they cannot be accessed and broken.

ImagePath implemented by substring of original path, and appending
the correct suffix. This needs to be improved as there's no check
if the last 4 path letters are the file extension. However, it will
do for now.