Implemented getShip in AOBAspect
This method gets data for a single ship - presumably selected in the
GUI, and saves it to memory as a ShipObject.

Renamed the XMLParser test method to PrepShipData and wrote a few XML
utilities to help with getting data.

Also, made some changes to the doc, but I've since forgotten
what these were.